Initialize domain_block_preprocessor module by extracting `preprocess_primary_block`
No logical changes.
