+++ domains/client/domain-executor/src/domain_block_preprocessor.rs
@@ -1,0 +1,100 @@
+use sc_client_api::BlockBackend;
+use sp_api::ProvideRuntimeApi;
+use sp_blockchain::HeaderBackend;
+use sp_domains::{DomainId, ExecutorApi, OpaqueBundles, SignedOpaqueBundles};
+use sp_runtime::generic::DigestItem;
+use sp_runtime::traits::{Block as BlockT, Header as HeaderT};
+use std::borrow::Cow;
+use subspace_core_primitives::Randomness;
+use subspace_wasm_tools::read_core_domain_runtime_blob;
+
+/// Domain-specific bundles extracted from the primary block.
+pub enum DomainBundles<Block, PBlock>
+where
+    Block: BlockT,
+    PBlock: BlockT,
+{
+    System(
+        OpaqueBundles<PBlock, Block::Hash>,
+        SignedOpaqueBundles<PBlock, Block::Hash>,
+    ),
+    Core(OpaqueBundles<PBlock, Block::Hash>),
+}
+
+pub type DomainBlockElements<Block, PBlock> = (
+    DomainBundles<Block, PBlock>,
+    Randomness,
+    Option<Cow<'static, [u8]>>,
+);
+
+/// Extracts the necessary materials for building a new domain block from the primary block.
+pub(crate) fn preprocess_primary_block<Block, PBlock, PClient>(
+    domain_id: DomainId,
+    primary_chain_client: &PClient,
+    block_hash: PBlock::Hash,
+) -> sp_blockchain::Result<DomainBlockElements<Block, PBlock>>
+where
+    Block: BlockT,
+    PBlock: BlockT,
+    PClient: HeaderBackend<PBlock> + BlockBackend<PBlock> + ProvideRuntimeApi<PBlock> + Send + Sync,
+    PClient::Api: ExecutorApi<PBlock, Block::Hash>,
+{
+    let extrinsics = primary_chain_client
+        .block_body(block_hash)?
+        .ok_or_else(|| {
+            sp_blockchain::Error::Backend(format!("BlockBody of {block_hash:?} unavailable"))
+        })?;
+
+    let header = primary_chain_client.header(block_hash)?.ok_or_else(|| {
+        sp_blockchain::Error::Backend(format!("BlockHeader of {block_hash:?} unavailable"))
+    })?;
+
+    let maybe_new_runtime = if header
+        .digest()
+        .logs
+        .iter()
+        .any(|item| *item == DigestItem::RuntimeEnvironmentUpdated)
+    {
+        let system_domain_runtime = primary_chain_client
+            .runtime_api()
+            .system_domain_wasm_bundle(block_hash)?;
+
+        let new_runtime = match domain_id {
+            DomainId::SYSTEM => system_domain_runtime,
+            DomainId::CORE_PAYMENTS => {
+                read_core_domain_runtime_blob(system_domain_runtime.as_ref(), domain_id)
+                    .map_err(|err| sp_blockchain::Error::Application(Box::new(err)))?
+                    .into()
+            }
+            _ => {
+                return Err(sp_blockchain::Error::Application(Box::from(format!(
+                    "No new runtime code for {domain_id:?}"
+                ))));
+            }
+        };
+
+        Some(new_runtime)
+    } else {
+        None
+    };
+
+    let shuffling_seed = primary_chain_client
+        .runtime_api()
+        .extrinsics_shuffling_seed(block_hash, header)?;
+
+    let domain_bundles = if domain_id.is_system() {
+        let (system_bundles, core_bundles) = primary_chain_client
+            .runtime_api()
+            .extract_system_bundles(block_hash, extrinsics)?;
+        DomainBundles::System(system_bundles, core_bundles)
+    } else if domain_id.is_core() {
+        let core_bundles = primary_chain_client
+            .runtime_api()
+            .extract_core_bundles(block_hash, extrinsics, domain_id)?;
+        DomainBundles::Core(core_bundles)
+    } else {
+        unreachable!("Open domains are unsupported")
+    };
+
+    Ok((domain_bundles, shuffling_seed, maybe_new_runtime))
+}
