+++ domains/client/domain-executor/src/domain_block_processor.rs
@@ -1,5 +1,5 @@
 use crate::fraud_proof::{find_trace_mismatch, FraudProofGenerator};
-use crate::utils::{shuffle_extrinsics, to_number_primitive, translate_number_type, DomainBundles};
+use crate::utils::{shuffle_extrinsics, to_number_primitive, translate_number_type};
 use crate::{ExecutionReceiptFor, TransactionFor};
 use codec::{Decode, Encode};
 use domain_block_builder::{BlockBuilder, BuiltBlock, RecordProof};
@@ -15,91 +15,11 @@ use sp_core::traits::CodeExecutor;
 use sp_domains::fraud_proof::FraudProof;
 use sp_domains::merkle_tree::MerkleTree;
 use sp_domains::{DomainId, ExecutionReceipt, ExecutorApi, OpaqueBundles};
-use sp_runtime::generic::DigestItem;
 use sp_runtime::traits::{Block as BlockT, HashFor, Header as HeaderT, One, Zero};
 use sp_runtime::Digest;
 use std::borrow::Cow;
 use std::sync::Arc;
 use subspace_core_primitives::Randomness;
-use subspace_wasm_tools::read_core_domain_runtime_blob;
-
-type DomainBlockElements<Block, PBlock> = (
-    DomainBundles<Block, PBlock>,
-    Randomness,
-    Option<Cow<'static, [u8]>>,
-);
-
-/// Extracts the necessary materials for building a new domain block from the primary block.
-pub(crate) fn preprocess_primary_block<Block, PBlock, PClient>(
-    domain_id: DomainId,
-    primary_chain_client: &PClient,
-    block_hash: PBlock::Hash,
-) -> sp_blockchain::Result<DomainBlockElements<Block, PBlock>>
-where
-    Block: BlockT,
-    PBlock: BlockT,
-    PClient: HeaderBackend<PBlock> + BlockBackend<PBlock> + ProvideRuntimeApi<PBlock> + Send + Sync,
-    PClient::Api: ExecutorApi<PBlock, Block::Hash>,
-{
-    let extrinsics = primary_chain_client
-        .block_body(block_hash)?
-        .ok_or_else(|| {
-            sp_blockchain::Error::Backend(format!("BlockBody of {block_hash:?} unavailable"))
-        })?;
-
-    let header = primary_chain_client.header(block_hash)?.ok_or_else(|| {
-        sp_blockchain::Error::Backend(format!("BlockHeader of {block_hash:?} unavailable"))
-    })?;
-
-    let maybe_new_runtime = if header
-        .digest()
-        .logs
-        .iter()
-        .any(|item| *item == DigestItem::RuntimeEnvironmentUpdated)
-    {
-        let system_domain_runtime = primary_chain_client
-            .runtime_api()
-            .system_domain_wasm_bundle(block_hash)?;
-
-        let new_runtime = match domain_id {
-            DomainId::SYSTEM => system_domain_runtime,
-            DomainId::CORE_PAYMENTS => {
-                read_core_domain_runtime_blob(system_domain_runtime.as_ref(), domain_id)
-                    .map_err(|err| sp_blockchain::Error::Application(Box::new(err)))?
-                    .into()
-            }
-            _ => {
-                return Err(sp_blockchain::Error::Application(Box::from(format!(
-                    "No new runtime code for {domain_id:?}"
-                ))));
-            }
-        };
-
-        Some(new_runtime)
-    } else {
-        None
-    };
-
-    let shuffling_seed = primary_chain_client
-        .runtime_api()
-        .extrinsics_shuffling_seed(block_hash, header)?;
-
-    let domain_bundles = if domain_id.is_system() {
-        let (system_bundles, core_bundles) = primary_chain_client
-            .runtime_api()
-            .extract_system_bundles(block_hash, extrinsics)?;
-        DomainBundles::System(system_bundles, core_bundles)
-    } else if domain_id.is_core() {
-        let core_bundles = primary_chain_client
-            .runtime_api()
-            .extract_core_bundles(block_hash, extrinsics, domain_id)?;
-        DomainBundles::Core(core_bundles)
-    } else {
-        unreachable!("Open domains are unsupported")
-    };
-
-    Ok((domain_bundles, shuffling_seed, maybe_new_runtime))
-}
 
 pub(crate) struct DomainBlockResult<Block, PBlock>
 where
