+++ domains/client/domain-executor/src/utils.rs
@@ -4,24 +4,11 @@ use rand::seq::SliceRandom;
 use rand::SeedableRng;
 use rand_chacha::ChaCha8Rng;
 use sp_consensus_slots::Slot;
-use sp_domains::{OpaqueBundles, SignedOpaqueBundles};
 use sp_runtime::traits::{Block as BlockT, NumberFor};
 use std::collections::{BTreeMap, VecDeque};
 use std::convert::TryInto;
 use std::fmt::Debug;
 use subspace_core_primitives::{Blake2b256Hash, BlockNumber, Randomness};
-
-pub(super) enum DomainBundles<Block, PBlock>
-where
-    Block: BlockT,
-    PBlock: BlockT,
-{
-    System(
-        OpaqueBundles<PBlock, Block::Hash>,
-        SignedOpaqueBundles<PBlock, Block::Hash>,
-    ),
-    Core(OpaqueBundles<PBlock, Block::Hash>),
-}
 
 /// Data required to produce bundles on executor node.
 #[derive(PartialEq, Clone, Debug)]
